bot/requirements.txt
python-telegram-bot>=21.0
qrcode[pil]>=7.4
python-dotenv>=1.0
Pillow>=10.0
